+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/computercraft/peripherals/EntityLinkPeripheral.kt
@@ -51,13 +51,7 @@ class EntityLinkPeripheral(private val blockEntity: EntityLinkBlockEntity) : Own
     }
 
     override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (other !is EntityLinkPeripheral) return false
-        if (!super.equals(other)) return false
-
-        if (blockEntity != other.blockEntity) return false
-
-        return true
+        return this === other
     }
 
     override fun hashCode(): Int {